Data conversion improved and bug fixed in it
This fixes a bug which can lead to interger overflow if data is close to
extrema. This has lead to incorporate the offset value in addition to the
scaling factor which in turn allows to convert unsigned integer to/from
signed integer.
bugfix: integer overflow when converting a data close to minimal value
bugfix: Data conversion intermediate is sometime badly chosen