Fix bug when channels are not all read
This fixes the bug found with files opened for reading. This bug occurred
when some channels were skip from being read. The reason was that the
feature of specifying -1 for an array index was actually not implemented!!!
A unit is now provided to test this feature as well as xdf_seek.