Add IEEE float rounding check to configure script
This add the check for making sure that the compiler will produce the
expected rounding behavior (especially when dealing with double). On x86,
the default FPU is x87 which use by default on linux and different platform
long double register internally. This would settings might screw up the
scaling computations and for sure will make testbdf fail.