Add test_seek in all format tests
The test consistes in generating a file with digital data in array whose
each couple (channel, tome offset) has a unique and predictible value. The
file is then reloaded, seeked at various position in the file (covering the
whole file range) and data is read and checked with the data prediction.