Add API xdf_writev()
This function is the same as xdf_read() but remplace the variable
argument list by an array of pointer. This allows to simplify the task users
of xdffileio who has a variable number of array. Those users were indeed
either forced to copy to one unique array or handle the case for each number
of input array they could use.