hard backup for gdf events in temporary files
In the gdf file format, events are store right at then end of the file.
Therefore it is impossible to stream those events the way the data is.

As a workaround, stream those datas into temporary files during the
acquisition. If all goes well, they are deleted during xdf_close()
Otherwise, they can be used to reconstruct a record in gdf format.

Change-Id: Idf6862b55533ce6ce5631869a7f8976eb03da8d3