create xdffile with RW for group and other
File use to be created with only RW permission for owner. This is a
rather annoying behavior which prevent easy file sharing. Hence we now
create new xdffile with RW for group and others as well which is the
usual permission of new file (given the usual umask the resulting
permission is R only for group and other).

Change-Id: I9831e4e24ba007f1a4f7a0e47c266d39d37d5e39