Update patch set 2
Patch Set 2:

I manually transformed warnings into errors within this script. Sphinx does not offer that specific option itself.
Not having a single function found raises a warning, this means that you need to have started the process of changing the comment format before this commit.
(This is why this was put after your commit)

Patch-set: 2