Update patch set 9
Patch Set 9:

> Could this be commit merged in https://review.gerrithub.io/c/mmlabs-mindmaze/xdffileio/+/489948?

As it is, it could if the doc commit is done first.

There is a difference between the two with regards to werror (only in autotools).
If no doc at all can be found, then autotools treats this as a warning (error).
This is why I put the commit after the doc commit from Marjorie.

We could also split this into two parts, the second being about transforming warnings into errors.

Patch-set: 9